requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
confluent-kafka==2.2.0
prometheus-client==0.19.0
numpy==1.24.0
pandas==2.1.0
scikit-learn==1.3.0
scipy==1.11.0
boto3==1.34.0
pyarrow==14.0.0
requests==2.31.0
python-multipart==0.0.6
pandera>=0.16.1
scipy>=1.11.0
pytest-cov>=4.1.0
codecov>=2.1.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0